+++ website/package.json
@@ -45,6 +45,7 @@
     "rehype-slug": "^6.0.0",
     "rehype-stringify": "^10.0.0",
     "remark-frontmatter": "^5.0.0",
+    "remark-gfm": "^4.0.0",
     "remark-heading-id": "^1.0.1",
     "remark-parse": "^11.0.0",
     "remark-rehype": "^11.0.0",
